requirements.txt
# For settting env
python-dotenv

# For price fetcher
web3
pandas

# For the model
pandas
numpy
scikit-learn
tensorflow
scipy
matplotlib
ta

# For the LLM
google-genai
